basira-backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9
sqlalchemy==2.0.30
aiosqlite==0.20.0
python-jose[cryptography]==3.3.0
bcrypt==4.0.1
openai==1.52.0
httpx==0.27.0
tiktoken==0.7.0
tenacity==8.3.0
sentence-transformers==3.0.1
torch==2.2.2
pinecone==3.2.2
PyMuPDF==1.24.11
pillow==10.3.0
requests==2.32.3
opencv-python-headless==4.9.0.80
numpy==1.26.4
python-dotenv==1.0.1
pydantic==2.7.1
pydantic-settings==2.3.0
cryptography==42.0.8
python-docx
fastapi-mail==1.4.1